requirements.txt
python-kucoin==2.1.3
pandas==1.0.3
websocket-client==1.4.2
websockets==10.4
